# feat: add stat function