--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,8 @@ module github.com/madlambda/benchcheck
 
 go 1.17
 
-require github.com/madlambda/spells v0.3.0 // indirect
+require (
+	github.com/google/go-cmp v0.5.7 // indirect
+	github.com/madlambda/spells v0.3.0 // indirect
+	golang.org/x/perf v0.0.0-20220411212318-84e58bfe0a7e // indirect
+)