The function encapsulates the heavy lifting done by benchstat, tried to use as little as possible from it, using default configurations + focusing on the simple scenario of comparing 2 benchmark run sets (benchstat is way more general/flexible).

Trying to capture a simple version of this:

```
benchstat old.txt new.txt
name        old time/op    new time/op    delta
GobEncode     13.6ms ± 1%    11.8ms ± 1%  -13.31%  (p=0.016 n=4+5)
JSONEncode    32.1ms ± 1%    31.8ms ± 1%     ~     (p=0.286 n=4+5)

name        old speed      new speed      delta
GobEncode   56.4MB/s ± 1%  65.1MB/s ± 1%  +15.36%  (p=0.016 n=4+5)
JSONEncode  60.4MB/s ± 1%  61.1MB/s ± 2%     ~     (p=0.286 n=4+5)
```

We can use the delta to compare and fail if it misses the target for a given specific metric (time/space/speed/etc).
After this building block we should be able to finish the tool, we can run benchs, get results, and call this function to provide statistics.