@i4ki 2022-05-01
I don't if make sense, but I had to add the command output to the CmdError.Error() in order to understand why my tests were failing:

```
    benchcheck_test.go:173: unexpected error[fail to run: /usr/local/go/bin/go test -bench=. ./... on dir /home/i4k/go/pkg/mod/github.com/!neoway!labs/gis@v0.0.0-20170702081752-562f220cd7ce: exit status 1 (output --- FAIL: TestEncodeFromCSV (0.00s)
            geohash_test.go:87: open /src/github.com/NeowayLabs/gis/encoding/geohash/_testdata/geohash.csv: no such file or directory
        FAIL
        exit status 1
```

```
diff --git a/benchcheck.go b/benchcheck.go
index 2b3bc00..279f10f 100644
--- a/benchcheck.go
+++ b/benchcheck.go
@@ -48,10 +48,11 @@ type CmdError struct {
 // Error returns the string representation of the error.
 func (c *CmdError) Error() string {
        return fmt.Sprintf(
-               "fail to run: %v on dir %s: %v",
+               "fail to run: %v on dir %s: %v (output %s)",
                c.Cmd,
                c.Cmd.Dir,
                c.Err,
+               c.Output,
        )
 }
```