@katcipis 2022-05-01
@i4ki for showing the cmd output of errors n tests, what about something like this ?

https://github.com/madlambda/benchcheck/pull/8

Example of how the error looks now:

https://github.com/madlambda/benchcheck/runs/6247375851?check_suite_focus=true#step:5:9