@i4ki 2022-05-01
old:
```
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
BenchmarkEncode-12         	11406358	       100.2 ns/op
BenchmarkStripe-12         	1000000000	         0.2726 ns/op
BenchmarkInterleaves-12    	303347204	         3.969 ns/op
PASS
```

new:
```
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
BenchmarkEncode-12              	18098190	        61.79 ns/op
BenchmarkEncodeFixedPoint-12    	30849720	        35.11 ns/op
BenchmarkDecode-12              	70118866	        16.54 ns/op
BenchmarkStripe-12              	1000000000	         0.2554 ns/op
BenchmarkInterleaves-12         	303250658	         3.949 ns/op
PASS
ok  	_/home/i4k/go/pkg/mod/github.com/_neoway_labs/gis@v0.0.0-20171205183537-0b758e536cc5/encoding/geohash	5.380s
```
results from my old geohash library (first commit versus optimized in the master).

just sharing maybe to use as testcase.